requirements.txt
# be sure to update the constraints file to match
# see https://airflow.apache.org/docs/apache-airflow/stable/installation/installing-from-pypi.html#constraints-files
apache-airflow==2.6.3